+++ src/api/tests/conftest.py
@@ -5,6 +5,7 @@
 from qualicharge.auth.factories import IDTokenFactory
 
 from .fixtures.app import client, client_auth
+from .fixtures.cli import runner
 from .fixtures.db import (
     db_engine,
     db_session,